[Gold III] Title: 세 용액, Time: 448 ms, Memory: 26736 KB -BaekjoonHub

+++ 백준/Gold/2473. 세 용액/세 용액.java
@@ -4,39 +4,39 @@ import java.util.Scanner;
 public class Main {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		long N = sc.nextLong();
-		long[] arr = new long[(int) N];
+		int N = sc.nextInt();
+		long[] arr = new long[N];
 		
-		for(long i = 0 ; i <N;i++) {
-			arr[(int) i] = sc.nextInt();
+		for(int i = 0 ; i <N;i++) {
+			arr[i] = sc.nextInt();
 		}
 		long[] temp = new long[3];
 		Arrays.sort(arr);
 		long answer=3_000_000_000L;
 		for(int i = 0 ;i<N-2;i++) {
-			long idx = i;
-			long L = idx+1;
-			long R = arr.length-1;
+			int idx = i;
+			int L = idx+1;
+			int R = arr.length-1;
 			while(L<R) {
-				long mid = arr[(int) idx]+arr[(int) L]+arr[(int) R];
+				long mid = arr[idx]+arr[L]+arr[R];
 				if(mid==0) {
-					System.out.println(arr[(int) idx]+" "+arr[(int) L]+" "+arr[(int) R]);
+					System.out.println(arr[idx]+" "+arr[L]+" "+arr[R]);
 					return;
 				}
 				if(mid>0) {
-					if(answer>Math.abs(arr[(int) idx]+arr[(int) L]+arr[(int) R])) {
-						answer=Math.abs(arr[(int) idx]+arr[(int) L]+arr[(int) R]);
-						temp[0]=arr[(int) idx];
-						temp[1]=arr[(int) L];
-						temp[2]=arr[(int) R];
+					if(answer>Math.abs(arr[idx]+arr[L]+arr[R])) {
+						answer=Math.abs(arr[idx]+arr[L]+arr[R]);
+						temp[0]=arr[idx];
+						temp[1]=arr[L];
+						temp[2]=arr[R];
 					}
 					R--;
 				}else {
-					if(answer>Math.abs(arr[(int) idx]+arr[(int) L]+arr[(int) R])) {
-						answer=Math.abs(arr[(int) idx]+arr[(int) L]+arr[(int) R]);
-						temp[0]=arr[(int) idx];
-						temp[1]=arr[(int) L];
-						temp[2]=arr[(int) R];
+					if(answer>Math.abs(arr[idx]+arr[L]+arr[R])) {
+						answer=Math.abs(arr[idx]+arr[L]+arr[R]);
+						temp[0]=arr[idx];
+						temp[1]=arr[L];
+						temp[2]=arr[R];
 					}
 					L++;
 				}
